+++ ios/Plugin/PhotoSlider/SliderViewController.swift
@@ -9,6 +9,14 @@ import UIKit
 
 import SDWebImage
 
+// swiftlint:disable file_length
+
+extension NSNotification.Name {
+    static var movieCompleted: Notification.Name {
+        return .init(rawValue:"movieCompleted")}
+}
+
+// swiftlint:disable type_body_length
 class SliderViewController: UIViewController {
 
     private var _numCells: CGFloat = 3
@@ -18,11 +26,13 @@ class SliderViewController: UIViewController {
     private var _position: IndexPath = [0, 0]
     private var _selectedPosition: IndexPath?
     private var _isShare: Bool = true
+    private var _isFilm: Bool = false
     private var _toast: Toast = Toast()
     private var _imageViewer: ImageScrollViewController?
     private var _maxZoomScale: Double = 3
     private var _compressionQuality: Double = 0.8
-
+    private var _movieOptions: [String: Any] = [:]
+    private var _movieObserver: Any?
     // MARK: - Set-up position
 
     var position: IndexPath {
@@ -65,20 +75,36 @@ class SliderViewController: UIViewController {
                 }
             }
             if self._options.keys.contains("compressionquality") {
-                if let compressionQuality = self._options[""] as? Double {
+                if let compressionQuality = self._options["compressionquality"]
+                    as? Double {
                     self._compressionQuality = compressionQuality
+                }
+            }
+            if self._options.keys.contains("movieoptions") {
+                if let movieOptions: [String: Any] = self._options["movieoptions"]
+                    as? [String: Any] {
+                    self._movieOptions = movieOptions
+                    self._isFilm = true
                 }
             }
         }
     }
+
+    // MARK: - Set-up Navigation Items
+
     lazy var navBar: UINavigationBar = { () -> UINavigationBar in
 
-        let navigationBar = UINavigationBar(frame: CGRect(x: 0, y: 35, width: view.frame.size.width, height: 64))
+        let navigationBar = UINavigationBar(
+            frame: CGRect(x: 0, y: 35, width: view.frame.size.width, height: 64))
         navigationBar.isTranslucent = true
         navigationBar.setBackgroundImage(UIImage(), for: .default)
         navigationBar.shadowImage = UIImage()
         let navigationItem = UINavigationItem()
-        navigationItem.rightBarButtonItem = mClose
+        if self._isFilm {
+            navigationItem.rightBarButtonItems = [mClose, mFilm]
+        } else {
+            navigationItem.rightBarButtonItem = mClose
+        }
         if self._isShare {
             navigationItem.leftBarButtonItem = mShare
         }
@@ -98,9 +124,9 @@ class SliderViewController: UIViewController {
             bClose.title = "Close"
             let fontSize: CGFloat = 18
             let font: UIFont = UIFont.boldSystemFont(ofSize: fontSize)
-            bClose.setTitleTextAttributes([NSAttributedString.Key
-                                            .foregroundColor: UIColor.white,
-                                           NSAttributedString.Key.font: font], for: .normal)
+            bClose.setTitleTextAttributes(
+                [NSAttributedString.Key.foregroundColor: UIColor.white,
+                 NSAttributedString.Key.font: font], for: .normal)
         }
         bClose.tintColor = .white
         bClose.action = #selector(closeButtonTapped)
@@ -118,9 +144,9 @@ class SliderViewController: UIViewController {
             bShare.title = "Share"
             let fontSize: CGFloat = 18
             let font: UIFont = UIFont.boldSystemFont(ofSize: fontSize)
-            bShare.setTitleTextAttributes([NSAttributedString.Key
-                                            .foregroundColor: UIColor.white,
-                                           NSAttributedString.Key.font: font], for: .normal)
+            bShare.setTitleTextAttributes(
+                [NSAttributedString.Key.foregroundColor: UIColor.white,
+                 NSAttributedString.Key.font: font], for: .normal)
         }
         bShare.tintColor = .white
         bShare.action = #selector(shareButtonTapped)
@@ -128,12 +154,39 @@ class SliderViewController: UIViewController {
 
     }()
 
+    lazy var  mFilm: UIBarButtonItem = {
+        let bFilm = UIBarButtonItem()
+        let image: UIImage?
+        if #available(iOS 13, *) {
+            let configuration = UIImage.SymbolConfiguration(scale: .large)
+            image = UIImage(systemName: "film",
+                            withConfiguration: configuration)
+            bFilm.image = image
+        } else {
+            bFilm.title = "Film"
+            let fontSize: CGFloat = 18
+            let font: UIFont = UIFont.boldSystemFont(ofSize: fontSize)
+            bFilm.setTitleTextAttributes(
+                [NSAttributedString.Key.foregroundColor: UIColor.white,
+                 NSAttributedString.Key.font: font], for: .normal)
+        }
+        bFilm.tintColor = .white
+        bFilm.action = #selector(filmButtonTapped)
+        return bFilm
+
+    }()
+
+    // MARK: - Set-up Layout
+
     lazy var layout: UICollectionViewFlowLayout = { () -> UICollectionViewFlowLayout in
         let mLayout: UICollectionViewFlowLayout = UICollectionViewFlowLayout()
         mLayout.minimumLineSpacing = 0
         mLayout.minimumInteritemSpacing = 0
         return mLayout
     }()
+
+    // MARK: - Set-up collectionView
+
     lazy var collectionView: UICollectionView = {
         let mColView = UICollectionView(
             frame: .zero, collectionViewLayout: layout)
@@ -144,6 +197,15 @@ class SliderViewController: UIViewController {
         return mColView
     }()
 
+    // MARK: - Deinit
+
+    deinit {
+        NotificationCenter.default
+            .removeObserver(self._movieObserver as Any)
+    }
+
+    // MARK: - loadView
+
     override func loadView() {
         view = UIView()
         let screenSize: CGRect = UIScreen.main.bounds
@@ -151,17 +213,26 @@ class SliderViewController: UIViewController {
         view.frame.size.height = screenSize.height
         view.sizeToFit()
         view.backgroundColor = .black
+        self._movieObserver = NotificationCenter.default
+            .addObserver(forName: .movieCompleted, object: nil,
+                         queue: nil, using: movieCompleted)
     }
+
+    // MARK: - viewDidLoad
 
     override func viewDidLoad() {
         super.viewDidLoad()
     }
+
+    // MARK: - viewWillAppear
 
     override func viewWillAppear(_ animated: Bool) {
         super.viewWillAppear(animated)
         view.addSubview(collectionView)
         view.addSubview(navBar)
     }
+
+    // MARK: - viewWillLayoutSubviews
 
     override func viewWillLayoutSubviews() {
         super.viewWillLayoutSubviews()
@@ -170,6 +241,9 @@ class SliderViewController: UIViewController {
         self.navBar.frame = CGRect(x: 0, y: 35,
                                    width: view.frame.size.width, height: 64)
     }
+
+    // MARK: - updateCollectionView
+
     private func updateCollectionView() {
         collectionView.translatesAutoresizingMaskIntoConstraints = false
         collectionView.topAnchor
@@ -196,6 +270,9 @@ class SliderViewController: UIViewController {
         }
 
     }
+
+    // MARK: - updateLayout
+
     private func updateLayout(_ size: CGSize) {
         let cellWidth = size.width
         let cellHeight = size.height
@@ -203,6 +280,8 @@ class SliderViewController: UIViewController {
         layout.itemSize = CGSize(width: cellWidth,
                                  height: cellHeight)
     }
+
+    // MARK: - viewWillTransition
 
     override func viewWillTransition(
         to size: CGSize,
@@ -214,9 +293,15 @@ class SliderViewController: UIViewController {
                                    width: view.frame.size.width, height: 64)
 
     }
+
+    // MARK: - closeButtonTapped
+
     @objc func closeButtonTapped() {
         self.dismiss(animated: true, completion: nil)
     }
+
+    // MARK: - shareButtonTapped
+
     @objc func shareButtonTapped() {
         var mPosition = self.position
         if let selPos = self._selectedPosition {
@@ -226,7 +311,8 @@ class SliderViewController: UIViewController {
             let imageView = UIImageView()
             imageView.contentMode = .scaleAspectFit
             imageView.clipsToBounds = true
-            imageView.sd_setImage(with: URL(string: imageUrl), placeholderImage: nil)
+            imageView.sd_setImage(with: URL(string: imageUrl),
+                                  placeholderImage: nil)
             if let image = imageView.image {
                 if let data = image.jpegData(compressionQuality:
                                                 CGFloat(_compressionQuality)) {
@@ -242,7 +328,43 @@ class SliderViewController: UIViewController {
         }
     }
 
+    // MARK: - filmButtonTapped
+
+    @objc func filmButtonTapped() {
+        let imagesToVideo: ImagesToVideo = ImagesToVideo()
+        imagesToVideo.imageList = self.imageList
+        imagesToVideo.options = self._movieOptions
+
+        //        DispatchQueue.global(qos: .userInitiated).async { [] in
+        imagesToVideo.createFilm()
+        //        }
+
+    }
+
+    // MARK: - Handle Notifications
+
+    @objc func movieCompleted(notification: Notification) {
+        guard let info = notification.userInfo as? [String: Any]
+        else { return }
+        DispatchQueue.main.async { [self] in
+            if let res = info["result"] as? Bool {
+                if res {
+                    let msg = "Movie has been created"
+                    _toast.showToast(view: self.view, message: msg,
+                                     font: .systemFont(ofSize: 16.0))
+                } else {
+                    let msg = "Movie creation failed"
+                    _toast.showToast(view: self.view, message: msg,
+                                     font: .systemFont(ofSize: 16.0))
+                }
+            }
+        }
+    }
 }
+// swiftlint:enable type_body_length
+
+// MARK: - Extensions
+
 extension SliderViewController: UICollectionViewDataSource {
 
     func numberOfSections(in collectionView: UICollectionView) -> Int {
@@ -256,7 +378,8 @@ extension SliderViewController: UICollectionViewDataSource {
     }
 
     func collectionView(_ collectionView: UICollectionView,
-                        cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
+                        cellForItemAt indexPath: IndexPath)
+    -> UICollectionViewCell {
 
         let cell = collectionView
             .dequeueReusableCell(withReuseIdentifier: SliderViewCell
@@ -339,3 +462,4 @@ extension UICollectionViewFlowLayout {
         return true
     }
 }
+// swiftlint:enable file_length