+++ ios/Plugin/PhotoViewer.swift
@@ -7,13 +7,20 @@ enum PhotoViewerError: Error {
 @objc public class PhotoViewer: NSObject {
     var collectionViewController: CollectionViewController?
 
+    // MARK: viewController
+
     @objc var viewController: CollectionViewController? {
         return collectionViewController
     }
 
+    // MARK: echo
+
     @objc public func echo(_ value: String) -> String {
         return value
     }
+
+    // MARK: show
+
     @objc public func show(_ imageList: [[String: String]],
                            options: [String: Any]) -> Bool {